Remove unnecessary export
In our production app we have a `script/dump_heap` bash script that does
some subshelling and calls out to a Ruby script to handle the
`ObjectSpace` bits and calculating the next file name based on existing
dump files. When I simplified things for the blurgh post, I forgot to
remove this `export`.

🎩 Thanks @jasonkarns for catching it.